Volume and Expression CC Now Supported
Note Velocity is now directly influenced by both Volume and Expression Control Changes. This ensures that this update will still make newly converted MIDIs compatible with Newer Eden MIDI Player and ScratchPFA, but is not guaranteed. Sustain CC is planned but highly unlikely to be developed and most likely to be removed.